+++ src/main/java/com/shootoff/camera/cameratypes/OptiTrackCamera.java
@@ -246,7 +246,7 @@ public class OptiTrackCamera implements Camera {
 	@Override
 	public boolean decreaseExposure() {
 		final int curExp = getExposure();
-		final int newExp = (int) (curExp - (.1 * (double) curExp));
+		final int newExp = (int) (curExp - (.1 * curExp));
 		logger.trace("curExp[ {} newExp {}", curExp, newExp);
 
 		if (newExp < 20) return false;
